fix: move PAGINATION_HEIGHT
