--- src/components/PaginationFooter/constants.ts
@@ -1,4 +1,0 @@
-const PAGINATION_HEIGHT = 40;
-
-// eslint-disable-next-line import/prefer-default-export
-export { PAGINATION_HEIGHT };
