+++ src/components/PaginationFooter/index.ts
@@ -1,2 +1,2 @@
+// eslint-disable-next-line import/prefer-default-export
 export { default as PaginationFooter } from './PaginationFooter';
-export { PAGINATION_HEIGHT } from './constants';
